Fix issues with and clean up ANNWrapper
If an entry to be merged had only company credited entry.get('credits', [])
would return that company entry without it being enclosed in a list.
This meant that the for loop would iterate over the key names in the
dictionary causing unwanted behavior. This has been fixed. Some
docstrings have been added and function names changed to be more PEP8
compliant.